refactor(blocktriple): split the text layer out of the core (#1334)
The last of the five internal cores. All of them are now stream-free:

    blocksignificand.hpp   25687   0 I/O
    blockfraction.hpp      25659   0 I/O
    blockbinary.hpp        33243   0 I/O
    ieee754_core.hpp       43486   0 I/O
    blocktriple.hpp        69791   0 I/O   <- was 98844 / 4

type_tag, the member to_string, to_binary, to_triple and the enum's string form
move to manipulators.hpp; operator<<, operator>> and the enum printer to
iostream.hpp. power_of_five moves too -- it is a decimal helper used only by
to_string, and keeping it in the class was what made the core carry
support/decimal.hpp. With it gone the core no longer needs decimal.hpp OR
native/integers.hpp.

ON THE 45000 TARGET: blocktriple sits above it at 69791, and that is a floor
rather than waste. It is essentially the sum of its two clean dependencies
(43486 native core + 25687 blocksignificand, partially overlapping). The target
was calibrated for a leaf header; for a composite the meaningful criterion is
0 I/O, which it meets.

THIS SPLIT HAS A FAILURE MODE THE OTHER FOUR DO NOT, and it is worth stating
because it will recur wherever a member moves out-of-line.

For free functions -- operator<<, to_binary -- deleting the friend declaration
turns a missing include into a COMPILE error, findable by the cheap sweep. That
is why the blockbinary, blocksignificand and blockfraction splits were verifiable
with -fsyntax-only.

to_string is a MEMBER. Its declaration has to stay in the class, so any call
compiles whether or not the definition is reachable, and the failure surfaces
only at LINK, only in targets that actually get built. The 1057-TU sweep passes
completely and proves nothing about it. Three link errors turned up this way,
each found by building:

    bt_teju      internal/blocktriple/conversion/teju.cpp
    posit_ulp    via number/posit/iostream.hpp and manipulators.hpp, both of
                 which call to_string on the blocktriple from to_value<>

The last two were fixed in the HEADERS rather than the TUs, which is what stops
the chase: one header fix took the sweep from 306 failures to 28.

THE abs() CHECK PAID OFF. Before extracting I classified every function in the
string-conversion region by return type and found abs() -- pure arithmetic
sitting inside the text region, exactly like urdiv and twosComplementFree before
it. Caught BEFORE the mistake this time rather than after. All eight text headers
across the five cores are now audited by return type and contain only text.

FOUR ORPHANED template<> LINES, and the fix that finally worked. Each extraction
left the template<...> header behind because the capture started at the
signature. I patched three one at a time, then wrote a general cleanup that was
too aggressive and deleted legitimate single-line forward declarations
(template<...> class blocktriple;). I reverted the file to main and redid the
extraction with the template header included in the capture from the start. Zero
orphans that time. Patching instances was the wrong move; fixing the extractor
was the right one.

Verified: the three headers and all 38 umbrellas compile on gcc 13.3, clang 18.1
and riscv64-linux-gnu cross; the 1057-TU sweep gives 17 failures, IDENTICAL to
the main baseline; CI_LITE builds and passes 456/456; the education and benchmark
trees build clean, since CI_LITE does not cover them and this PR's failure mode
is link-only; check-ascii clean; zero added lines over 120 columns.

Relates to #1334
