+++ internal/blocktriple/api/constexpr.cpp
@@ -14,6 +14,7 @@
 
 // minimum set of include files to reflect source code dependencies
 #include <universal/internal/blocktriple/blocktriple.hpp>
+#include <universal/internal/blocktriple/iostream.hpp>   // (#1334)
 
 namespace sw::experiment {
 	template<size_t nbits, typename bt = uint32_t>
