+++ internal/blocktriple/conversion/rounding.cpp
@@ -17,6 +17,7 @@
 #include <universal/internal/blocktriple/blocktriple.hpp>
 #include <universal/verification/test_suite.hpp>
 #include <universal/internal/blocksignificand/manipulators.hpp>   // (#1334)
+#include <universal/internal/blocktriple/manipulators.hpp>   // (#1334)
 
 namespace sw { namespace universal {
